[DEV] update

+++ app/src/main/java/com/khs/lovelynote/viewmodel/factory/BoardDetailVMFactory.kt
@@ -5,7 +5,7 @@ import androidx.lifecycle.ViewModel
 import androidx.lifecycle.ViewModelProvider
 import com.khs.lovelynote.viewmodel.BoardDetailVM
 
-class BoardDetailVMFactory(private val mApplication: Application, private val param1: Int) :
+class BoardDetailVMFactory(private val mApplication: Application, private val param1: Long) :
     ViewModelProvider.Factory {
     override fun <T : ViewModel?> create(modelClass: Class<T>): T {
         return BoardDetailVM(mApplication, param1) as T
